requirements-tests.txt
fabric
pymox
django-test-utilities>=0.5
model-mommy==1.2.6
nose
django-nose
south
tox==2.6.0